License popup integration with the shell
