+++ src/ServiceInsight/Shell/IShellView.cs
@@ -5,7 +5,5 @@
         void ChangeTheme(string name);
 
         void SelectTab(string name);
-
-        object LicenseStatus { get; }
     }
 }
