+++ src/ServiceInsight/Shell/LicensePopupView.xaml
@@ -4,9 +4,9 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006" 
              xmlns:d="http://schemas.microsoft.com/expression/blend/2008" 
              xmlns:controls="clr-namespace:ServiceInsight.Controls"
-             MouseLeftButtonUp="OnMouseLeftButtonUp"
+             xmlns:shell="clr-namespace:ServiceInsight.Shell"
              mc:Ignorable="d" 
-             d:DesignHeight="450" d:DesignWidth="800">
+             d:DesignHeight="450" d:DesignWidth="800" d:DataContext="{d:DesignInstance shell:LicensePopupViewModel}">
     <UserControl.Resources>
         <Style TargetType="{x:Type controls:LinkedIcon}">
             <Setter Property="Foreground" Value="{DynamicResource {x:Static SystemColors.HighlightBrushKey}}" />
@@ -72,10 +72,18 @@
                     <RowDefinition Height="Auto" />
                 </Grid.RowDefinitions>
 
-                <TextBlock Grid.Row="0" FontSize="14" Margin="18,18,18,5" 
-                                   FontWeight="Bold" Text="Warning: Platform license expiring in 3 days" />
-                <TextBlock Grid.Row="1" FontSize="14" Margin="18,0,18,5" 
-                                   TextWrapping="WrapWithOverflow" Text="Once the license expires you'll no longer be able to continue using the Particular Service Platform. This is just a bit longer text." />
+                <TextBlock Grid.Row="0" 
+                           FontSize="14"
+                           Margin="18,18,18,5" 
+                           FontWeight="Bold" 
+                           Foreground="Black"
+                           Text="Warning: Platform license expiring in 3 days" />
+                <TextBlock Grid.Row="1" 
+                           FontSize="14" 
+                           Margin="18,0,18,5" 
+                           Foreground="Black"
+                           TextWrapping="WrapWithOverflow" 
+                           Text="Once the license expires you'll no longer be able to continue using the Particular Service Platform. This is just a bit longer text." />
 
                 <Button Click="OnCloseButtonClicked" 
                         Style="{StaticResource CloseButton}"
@@ -91,22 +99,25 @@
                     </Grid.ColumnDefinitions>
 
                     <controls:LinkedIcon HorizontalAlignment="Left" 
-                                         Grid.Column="0" Margin="18,0,0,10" 
+                                         Grid.Column="0" 
+                                         Margin="18,0,0,10" 
                                          Command="{Binding ContactUs}"
-                                         Text="Contact us" ImageSource="/Images/Contact.png" />
+                                         Text="Contact us" 
+                                         ImageSource="/Images/Contact.png" />
 
                     <controls:LinkedIcon HorizontalAlignment="Right" 
-                                         Grid.Column="1" Margin="0,0,18,10" 
+                                         Grid.Column="1" 
+                                         Margin="0,0,18,10" 
                                          Command="{Binding ManageLicense}"
-                                         Text="Manage license..." ImageSource="/Images/License.png" />
-
+                                         Text="Manage license..." 
+                                         ImageSource="/Images/License.png" />
                 </Grid>
             </Grid>
         </Border>
 
         <Path x:Name="Container"
                       Grid.Row="1"
-                      Margin="20,-10,0,0"
+                      Margin="40,-10,0,0"
                       Data="M 0,0 L 10,20 20,0"
                       Fill="#d6d6d6">
         </Path>
