+++ src/ServiceInsight/Shell/Shell.xaml
@@ -52,7 +52,7 @@
                     AllowCustomization="True"
                     AllowUIAutomationSupport="True"
                     AutomationProperties.AutomationId="BarManager"
-                    dxc:DXSerializer.LayoutVersion="2"
+                    dxc:DXSerializer.LayoutVersion="3"
                     dxc:DXSerializer.SerializationID="BarManager">
         <!--  Categories  -->
         <dxb:BarManager.Categories>
@@ -202,13 +202,13 @@
 
                             <TextBlock VerticalAlignment="Center" Text="{Binding Path=StatusBarManager.LicenseStatusMessage, Mode=OneWay}" />
 
-                            <!--<Popup IsOpen="{Binding StatusBarManager.ShowLicensePopup}"
+                            <Popup IsOpen="{Binding StatusBarManager.ShowLicensePopup, Mode=OneWay}"
                                    StaysOpen="False"
                                    AllowsTransparency="True"
                                    HorizontalOffset="50" Placement="Top"
                                    PlacementTarget="{Binding ElementName=StatusPanel}">
-                                <local:LicensePopupView />
-                            </Popup>-->
+                                <local:LicensePopupView cal:Bind.Model="{Binding StatusBarManager}" />
+                            </Popup>
                         </StackPanel>
                     </DataTemplate>
                 </dxb:BarStaticItem.ContentTemplate>
@@ -281,9 +281,9 @@
                 <dxb:Bar.ItemLinks>
                     <dxb:BarStaticItemLink BarItemName="Status"  />
                     <dxb:BarItemLinkSeparator />
-                    <dxb:BarStaticItemLink BarItemName="LicenseStatusBar" />
-                    <dxb:BarItemLinkSeparator />
                     <dxb:BarStaticItemLink BarItemName="Registration" />
+                    <dxb:BarItemLinkSeparator />
+                    <dxb:BarStaticItemLink BarItemName="LicenseStatusBar" />
                 </dxb:Bar.ItemLinks>
             </dxb:Bar>
         </dxb:BarManager.Bars>
