+++ src/ServiceInsight/Shell/Shell.xaml.cs
@@ -53,8 +53,6 @@
             }
         }
 
-        public object LicenseStatus => LicenseStatusBar.Content;
-
         public void OnSaveLayout(ISettingsProvider settingProvider)
         {
             var layoutSetting = settingProvider.GetSettings<ShellLayoutSettings>();
